+++ crates/runtime-api/src/polkavm_imports.rs
@@ -22,6 +22,8 @@ pub static BALANCE_OF: &str = "balance_of";
 
 pub static BASE_FEE: &str = "base_fee";
 
+pub static BLOCK_AUTHOR: &str = "block_author";
+
 pub static BLOCK_HASH: &str = "block_hash";
 
 pub static BLOCK_NUMBER: &str = "block_number";
@@ -80,13 +82,14 @@ pub static WEIGHT_TO_FEE: &str = "weight_to_fee";
 
 /// All imported runtime API symbols.
 /// Useful for configuring common attributes and linkage.
-pub static IMPORTS: [&str; 34] = [
+pub static IMPORTS: [&str; 35] = [
     SBRK,
     MEMORY_SIZE,
     ADDRESS,
     BALANCE,
     BALANCE_OF,
     BASE_FEE,
+    BLOCK_AUTHOR,
     BLOCK_HASH,
     BLOCK_NUMBER,
     CALL,